Fix more MM builder issues (#3373)
* fix unpair of signed IDs

* Fix MM builder issues + invalid translation

* Remove disrupting line breaks for help page and inbox list

* Fix loop of mm parsing in /plot f set

* de-clutter help page

+++ Core/src/main/java/com/plotsquared/core/plot/PlotId.java
@@ -110,7 +110,7 @@ public final class PlotId {
      * @return Plot ID
      */
     public static @NonNull PlotId unpair(final int hash) {
-        return PlotId.of(hash >> 16, hash & 0xFFFF);
+        return PlotId.of(hash >> 16, (short) (hash & 0xFFFF));
     }
 
     /**
